Web scraper
This program scraps the data from the website and selects the data where the user wants to collect. This program in particular scraps the data from the website called "Hacker News". It scraps the news on the list depending on their voting and gets sorted when it gets displayed for the users.